event refactor
