+++ src/main/java/gov/nist/csd/pm/epp/EPP.java
@@ -55,7 +55,7 @@ public class EPP {
 
             List<Rule> rules = obligation.getRules();
             for(Rule rule : rules) {
-                if(!eventMatches(eventCtx.getUserCtx(), eventCtx.getEvent(), eventCtx.getTarget(), rule.getEventPattern())) {
+                if(!eventCtx.matchesPattern(rule.getEventPattern(), pap.getGraphAdmin())) {
                     continue;
                 }
 
@@ -113,167 +113,6 @@ public class EPP {
         }
 
         return true;
-    }
-
-    private boolean eventMatches(UserContext userCtx, String event, Node target, EventPattern match) throws PMException {
-        if(match.getOperations() != null &&
-                !match.getOperations().contains(event)) {
-            return false;
-        }
-
-        Subject matchSubject = match.getSubject();
-        PolicyClass matchPolicyClass = match.getPolicyClass();
-        Target matchTarget = match.getTarget();
-
-        return subjectMatches(userCtx.getUser(), userCtx.getProcess(), matchSubject) &&
-                pcMatches(userCtx.getUser(), userCtx.getProcess(), matchPolicyClass) &&
-                targetMatches(target, matchTarget);
-    }
-
-    private boolean subjectMatches(String user, String process, Subject matchSubject) throws PMException {
-        if(matchSubject == null) {
-            return true;
-        }
-
-        // any user
-        if((matchSubject.getAnyUser() == null && matchSubject.getUser() == null && matchSubject.getProcess() == null) ||
-                (matchSubject.getAnyUser() != null && matchSubject.getAnyUser().isEmpty())) {
-            return true;
-        }
-
-        // get the current user node
-        Node userNode = pap.getGraphAdmin().getNode(user);
-
-        if (checkAnyUser(userNode, matchSubject.getAnyUser())) {
-            return true;
-        }
-
-        if(matchSubject.getUser() != null && matchSubject.getUser().equals(userNode.getName())) {
-            return true;
-        }
-
-        return matchSubject.getProcess() != null &&
-                matchSubject.getProcess().getValue().equals(process);
-    }
-
-    private boolean checkAnyUser(Node userNode, List<String> anyUser) throws PMException {
-        if (anyUser == null || anyUser.isEmpty()) {
-            return true;
-        }
-
-        DepthFirstSearcher dfs = new DepthFirstSearcher(pap.getGraphAdmin());
-
-        // check each user in the anyUser list
-        // there can be users and user attributes
-        for (String u : anyUser) {
-            Node anyUserNode = pap.getGraphAdmin().getNode(u);
-
-            // if the node in anyUser == the user than return true
-            if (anyUserNode.getName().equals(userNode.getName())) {
-                return true;
-            }
-
-            // if the anyUser is not an UA, move to the next one
-            if (anyUserNode.getType() != UA) {
-                continue;
-            }
-
-            Set<String> nodes = new HashSet<>();
-            Visitor visitor = node -> {
-                if (node.getName().equals(userNode.getName())) {
-                    nodes.add(node.getName());
-                }
-            };
-            dfs.traverse(userNode, (c, p) -> {}, visitor, Direction.PARENTS);
-
-            if (nodes.contains(anyUserNode.getName())) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private boolean pcMatches(String user, String process, PolicyClass matchPolicyClass) {
-        // not yet implemented
-        return true;
-    }
-
-    private boolean targetMatches(Node target, Target matchTarget) throws PMException {
-        if(matchTarget == null) {
-            return true;
-        }
-
-        if (matchTarget.getPolicyElements() == null &&
-                matchTarget.getContainers() == null) {
-            return true;
-        }
-
-        if(matchTarget.getContainers() != null) {
-            if (matchTarget.getContainers().isEmpty()) {
-                return true;
-            }
-
-            // check that target is contained in any container
-            Set<Node> containers = getContainersOf(target.getName());
-            for(EvrNode evrContainer : matchTarget.getContainers()) {
-                for(Node contNode : containers) {
-                    if (nodesMatch(evrContainer, contNode)) {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        } else if(matchTarget.getPolicyElements() != null) {
-            if (matchTarget.getPolicyElements().isEmpty()) {
-                return true;
-            }
-
-            // check that target is in the list of policy elements
-            for(EvrNode evrNode : matchTarget.getPolicyElements()) {
-                if(nodesMatch(evrNode, target)) {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        return false;
-    }
-
-    private boolean nodesMatch(EvrNode evrNode, Node node) {
-        if(!evrNode.getName().equals(node.getName())) {
-            return false;
-        }
-
-        if(!evrNode.getType().equals(node.getType().toString())) {
-            return false;
-        }
-
-        for (String k : evrNode.getProperties().keySet()) {
-            String v = evrNode.getProperties().get(k);
-            if(node.getProperties().containsKey(k)) {
-                if(!node.getProperties().get(k).equals(v)) {
-                    return false;
-                }
-            } else {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private Set<Node> getContainersOf(String name) throws PMException {
-        Set<Node> nodes = new HashSet<>();
-        Set<String> parents = pap.getGraphAdmin().getParents(name);
-        for (String parent : parents) {
-            nodes.add(pap.getGraphAdmin().getNode(parent));
-            nodes.addAll(getContainersOf(parent));
-        }
-        return nodes;
     }
 
     private void applyAction(UserContext definingUser, String label, EventContext eventCtx, Action action) throws PMException {
