+++ src/main/java/gov/nist/csd/pm/pip/obligations/evr/EVRParser.java
@@ -1,8 +1,5 @@
 package gov.nist.csd.pm.pip.obligations.evr;
 
-import gov.nist.csd.pm.exceptions.PMException;
-import gov.nist.csd.pm.pip.graph.Graph;
-import gov.nist.csd.pm.pip.graph.model.nodes.Node;
 import gov.nist.csd.pm.pip.obligations.model.*;
 import gov.nist.csd.pm.pip.obligations.model.actions.*;
 import gov.nist.csd.pm.pip.obligations.model.functions.Arg;
@@ -12,14 +9,10 @@ import org.yaml.snakeyaml.Yaml;
 
 import java.io.InputStream;
 import java.util.*;
-import java.util.logging.Logger;
 
 public class EVRParser {
 
-    private static Logger logger = Logger.getLogger(EVRParser.class.getName());
-
-    private static <T> T getObject(Object o, Class<T> type) throws EVRException {
-        logger.info(String.format("getting object %s of type %s", o, type));
+    public static <T> T getObject(Object o, Class<T> type) throws EVRException {
         if(!type.isInstance(o)) {
             throw new EVRException("expected " + type + " got " + o.getClass() + " at \"" + o + "\"");
         }
@@ -27,7 +20,21 @@ public class EVRParser {
         return type.cast(o);
     }
 
-    public static Obligation parse(String user, String yml) throws EVRException {
+    private EventParser systemEventParser = new SystemEventParser();
+    private Map<String, EventParser> customEventParsers;
+
+    public EVRParser(EventParser ... customEventParsers) {
+        this.customEventParsers = new HashMap<>();
+        for (EventParser parser : customEventParsers) {
+            if (this.customEventParsers.containsKey(parser.key())) {
+                throw new IllegalArgumentException("EventParser duplicate key: " + parser.key());
+            }
+
+            this.customEventParsers.put(parser.key(), parser);
+        }
+    }
+
+    public Obligation parse(String user, String yml) throws EVRException {
         Yaml yaml = new Yaml();
         Map<Object, Object> map = yaml.load(yml);
 
@@ -38,14 +45,14 @@ public class EVRParser {
      * label: string required
      * rules: array
      */
-    public static Obligation parse(String user, InputStream is) throws EVRException {
+    public Obligation parse(String user, InputStream is) throws EVRException {
         Yaml yaml = new Yaml();
         Map<Object, Object> map = yaml.load(is);
 
         return parse(user, map);
     }
 
-    private static Obligation parse(String user, Map<Object, Object> map) throws EVRException {
+    private Obligation parse(String user, Map<Object, Object> map) throws EVRException {
         Obligation obligation = new Obligation(user);
 
         String label = getObject(map.get("label"), String.class);
@@ -73,7 +80,7 @@ public class EVRParser {
      * event: object required
      * response: object required
      */
-    public static Rule parseRule(Object o) throws EVRException {
+    public Rule parseRule(Object o) throws EVRException {
         if(!(o instanceof Map)) {
             throw new EVRException("rule should be a map, got " + o.getClass() + " in " + o);
         }
@@ -102,244 +109,39 @@ public class EVRParser {
         return rule;
     }
 
-    /**
-     * subject: object if omitted any subject matches
-     * policyClass: object if omitted any policyClass matches
-     * operations: array if omitted any operation matches
-     * target: object if omitted any target matches
-     */
-    protected static EventPattern parseEvent(Object o) throws EVRException {
-        if(!(o instanceof Map)) {
-            throw new EVRException("event should be a Map, got " + o.getClass() + " in " + o);
-        }
-
-        Map map = (Map)o;
-        EventPattern eventPattern = new EventPattern();
-        if(map.containsKey("subject")) {
-            Subject subject = parseSubject(map.get("subject"));
-            eventPattern.setSubject(subject);
-        }
-
-        if(map.containsKey("policyClass")) {
-            PolicyClass policyClass = parsePolicyClass(map.get("policyClass"));
-            eventPattern.setPolicyClass(policyClass);
-        }
-
-        if(map.containsKey("operations")) {
-            List<String> operations = parseOperations(map.get("operations"));
-            eventPattern.setOperations(operations);
-        }
-
-        if(map.containsKey("target")) {
-            Target target = parseTarget(map.get("target"));
-            eventPattern.setTarget(target);
-        }
-
-        return eventPattern;
-    }
-
-    /**
-     * subject:
-     *   user:
-     *   anyUser:
-     *   process:
-     */
-    protected static Subject parseSubject(Object o) throws EVRException {
-        // a null map means any subject
-        if(o == null) {
-            return new Subject();
-        } else if(!(o instanceof Map)) {
-            throw new EVRException("event subject should be a Map, got " + o.getClass());
-        }
-
-        Map map = (Map)o;
-        if(map.size() != 1) {
-            throw new EVRException("only one element is expected for an event subject, got " + map);
-        }
-
-        if(map.containsKey("user")) {
-            return parseSubjectUser(map.get("user"));
-        } else if(map.containsKey("anyUser")) {
-            return parseSubjectAnyUser(map.get("anyUser"));
-        } else if(map.containsKey("process")) {
-            return parseProcess(map.get("process"));
-        }
-
-        throw new EVRException("invalid subject specification " + map);
-    }
-
-    /**
-     * 1. A specific policy element
-     *    policyElements:
-     *      - name: name
-     *        type: type
-     *
-     * 2. Any policy element
-     *    policyElements: or omit
-     *
-     * 3. Any policy element that is contained in other policy elements
-     *    policyElements: or omit
-     *    containers:
-     *      -
-     *      -
-     *
-     * 4. Any policy element from a set of policy elements
-     *    policyElements:
-     *      - name: name
-     *        type: type
-     *      - name: name
-     *        type: type
-     *
-     * policyElements: array if omitted any policy element will match
-     * ---
-     * containers: array if omitted any container will match
-     *
-     * ony one of containers or policyElements is allowed
-     * if both are omitted it will be "any policyElement in any container"
-     * if containers is present then it will be "any policyElement in the containers",
-     *   regardless of if policyElements is present
-     * if policyElements is present its "any policyElement from the list provided"
-     */
-    protected static Target parseTarget(Object o) throws EVRException {
-        Target target = new Target();
-        if(o == null) {
-            return target;
-        }
-
+    private EventPattern parseEvent(Object o) throws EVRException {
+        // this map should be
+        //   1. system event
+        //   2. custom event with the format: key -> custom tags
         Map map = getObject(o, Map.class);
-        if(map.containsKey("containers")){
-            target.setContainers(parseContainers(getObject(map.get("containers"), List.class)));
-        } else if (map.containsKey("policyElements")) {
-            target.setPolicyElements(parsePolicyElements(getObject(map.get("policyElements"), List.class)));
+        if (!isSystemEvent(map)) {
+            return parseCustomEvent(map);
         }
 
-        return target;
+        return systemEventParser.parse(map);
     }
 
-    private static List<EvrNode> parseContainers(List contList) throws EVRException {
-        List<EvrNode> containers = new ArrayList<>();
-        for(Object conObj : contList) {
-            Map conMap = getObject(conObj, Map.class);
-            String name = getObject(conMap.get("name"), String.class);
-            String type = getObject(conMap.get("type"), String.class);
-            Map propsMap = new HashMap();
-            if(conMap.containsKey("properties")) {
-                propsMap = getObject(conMap.get("properties"), Map.class);
-            }
-            Map<String, String> properties = new HashMap<>();
-            for(Object propObj : propsMap.keySet()) {
-                String value = getObject(propsMap.get(propObj), String.class);
-                properties.put((String) propObj, value);
-            }
-
-            containers.add(new EvrNode(name, type, properties));
+    private EventPattern parseCustomEvent(Map map) throws EVRException {
+        // the given map should only have 1 key
+        if (map.size() != 1) {
+            throw new EVRException("only one key expected in custom event: " + map);
         }
 
-        return containers;
+        String key = (String) map.keySet().iterator().next();
+
+        if (!this.customEventParsers.containsKey(key)) {
+            throw new EVRException("unregistered custom event key: " + key);
+        }
+
+        EventParser parser = this.customEventParsers.get(key);
+        return parser.parse(getObject(map.get(key), Map.class));
     }
 
-    /**
-     * policyElements:
-     *   - name:
-     *     type:
-     *   - name:
-     *     type:
-     */
-    private static List<EvrNode> parsePolicyElements(List list) throws EVRException {
-        List<EvrNode> policyElements = new ArrayList<>();
-
-        if(list == null) {
-            return policyElements;
-        }
-
-        // check that each element in the array is a string
-        for(Object l : list) {
-            policyElements.add(parseEvrNode(getObject(l, Map.class)));
-        }
-
-        return policyElements;
-    }
-
-    /**
-     * operations:
-     *   - ""
-     *   - ""
-     */
-    protected static List<String> parseOperations(Object o) throws EVRException {
-        if(o == null) {
-            return new ArrayList<>();
-        }
-
-        List opsList = getObject(o, List.class);
-        List<String> operations = new ArrayList<>();
-        for(Object op : opsList) {
-            operations.add(getObject(op, String.class));
-        }
-
-        return operations;
-    }
-
-    /**
-     * policyClass:
-     *   anyOf:
-     *   ---
-     *   eachOf:
-     *
-     * One of anyOf/eachOf is allowed, both are arrays of string
-     */
-    protected static PolicyClass parsePolicyClass(Object o) throws EVRException {
-        PolicyClass policyClass = new PolicyClass();
-        if(o == null) {
-            return policyClass;
-        }
-
-        Map map = (Map)o;
-        if(map.size() > 1) {
-            throw new EVRException("expected one of (anyOf, eachOf), got " + map.keySet());
-        }
-
-        if(map.containsKey("anyOf")) {
-            List<String> pcs = new ArrayList<>();
-            List list = getObject(map.get("anyOf"), List.class);
-            for(Object obj : list) {
-                pcs.add((String) obj);
-            }
-            policyClass.setAnyOf(pcs);
-        } else if(map.containsKey("eachOf")) {
-            List<String> pcs = new ArrayList<>();
-            List list = getObject(map.get("eachOf"), List.class);
-            for(Object obj : list) {
-                pcs.add((String) obj);
-            }
-            policyClass.setEachOf(pcs);
-        }
-
-        return policyClass;
-    }
-
-
-    private static Subject parseProcess(Object o) throws EVRException {
-        String process = getObject(o, String.class);
-        return new Subject(new EvrProcess(process));
-    }
-
-    private static Subject parseSubjectAnyUser(Object o) throws EVRException {
-        List<String> anyUser = new ArrayList<>();
-        Subject subject = new Subject(anyUser);
-        if (o == null) {
-            return subject;
-        }
-
-        List list = getObject(o, List.class);
-        for(Object obj : list) {
-            anyUser.add(getObject(obj, String.class));
-        }
-
-        return subject;
-    }
-
-    private static Subject parseSubjectUser(Object o) throws EVRException {
-        return new Subject(getObject(o, String.class));
+    private boolean isSystemEvent(Map map) {
+        return map.containsKey("subject") ||
+                map.containsKey("operations") ||
+                map.containsKey("policyClass") ||
+                map.containsKey("target");
     }
 
     /**
@@ -347,7 +149,7 @@ public class EVRParser {
      *   condition:
      *   actions:
      */
-    protected static ResponsePattern parseResponse(Object o) throws EVRException {
+    protected ResponsePattern parseResponse(Object o) throws EVRException {
         ResponsePattern responsePattern = new ResponsePattern();
         if(o == null) {
             return responsePattern;
@@ -379,7 +181,7 @@ public class EVRParser {
      *
      *   if all functions evaluate to true the condition is true
      */
-    private static Condition parseCondition(Object o) throws EVRException {
+    private Condition parseCondition(Object o) throws EVRException {
         List list = getObject(o, List.class);
         List<Function> functions = new ArrayList<>();
         for(Object l : list) {
@@ -400,7 +202,7 @@ public class EVRParser {
      *
      *   if all functions evaluate to false the condition is true
      */
-    private static NegatedCondition parseNegatedCondition(Object o) throws EVRException {
+    private NegatedCondition parseNegatedCondition(Object o) throws EVRException {
         List list = getObject(o, List.class);
         List<Function> functions = new ArrayList<>();
         for(Object l : list) {
@@ -414,7 +216,7 @@ public class EVRParser {
         return negatedCondition;
     }
 
-    private static Action parseAction(Map map) throws EVRException {
+    private Action parseAction(Map map) throws EVRException {
         Condition condition = null;
         if (map.containsKey("condition")) {
             condition = parseCondition(map.get("condition"));
@@ -455,12 +257,12 @@ public class EVRParser {
         return action;
     }
 
-    private static Action parseFunctionAction(Object o) throws EVRException {
+    private Action parseFunctionAction(Object o) throws EVRException {
         Map map = getObject(o, Map.class);
         return new FunctionAction(parseFunction(map));
     }
 
-    private static EvrNode parseEvrNode(Map map) throws EVRException {
+    public static EvrNode parseEvrNode(Map map) throws EVRException {
         if(map == null) {
             throw new EVRException("null EVR node found");
         }
@@ -508,7 +310,7 @@ public class EVRParser {
      *   ---
      *   - rule:
      */
-    private static Action parseCreateAction(Object o) throws EVRException {
+    private Action parseCreateAction(Object o) throws EVRException {
         if(o == null) {
             throw new EVRException("create action cannot be null or empty");
         }
@@ -543,7 +345,7 @@ public class EVRParser {
      *
      * the "like" element described in the standard is not implemented
      */
-    private static Action parseAssignAction(Object o) throws EVRException {
+    private Action parseAssignAction(Object o) throws EVRException {
         if(o == null) {
             throw new EVRException("assign action cannot be null or empty");
         }
@@ -585,7 +387,7 @@ public class EVRParser {
      *     function:
      *
      */
-    private static Action parseGrantAction(Object o) throws EVRException {
+    private Action parseGrantAction(Object o) throws EVRException {
         if(o == null) {
             throw new EVRException("grant action cannot be null or empty");
         }
@@ -619,11 +421,11 @@ public class EVRParser {
         return action;
     }
 
-    private static EvrNode parseSubject(Map map) throws EVRException {
+    private EvrNode parseSubject(Map map) throws EVRException {
         return parseEvrNode(map);
     }
 
-    private static EvrNode parseTarget(Map map) throws EVRException {
+    private EvrNode parseTarget(Map map) throws EVRException {
         return parseEvrNode(map);
     }
 
@@ -652,7 +454,7 @@ public class EVRParser {
      *         complement: true|false
      *
      */
-    private static Action parseDenyAction(Object o) throws EVRException {
+    private Action parseDenyAction(Object o) throws EVRException {
         if(o == null) {
             throw new EVRException("deny action cannot be null or empty");
         }
@@ -690,7 +492,7 @@ public class EVRParser {
         return action;
     }
 
-    private static DenyAction.Target parseDenyActionTarget(Object o) throws EVRException {
+    private DenyAction.Target parseDenyActionTarget(Object o) throws EVRException {
         Map targetMap = getObject(o, Map.class);
 
         DenyAction.Target target = new DenyAction.Target();
@@ -718,7 +520,7 @@ public class EVRParser {
         return target;
     }
 
-    private static List<DenyAction.Target.Container> parseDenyActionTargetContainers(List list) throws EVRException {
+    private List<DenyAction.Target.Container> parseDenyActionTargetContainers(List list) throws EVRException {
         List<DenyAction.Target.Container> containers = new ArrayList<>();
         for(Object contObj : list) {
             Map contMap = getObject(contObj, Map.class);
@@ -780,7 +582,7 @@ public class EVRParser {
      *   rules:
      *     - ""
      */
-    private static Action parseDeleteAction(Object o) throws EVRException {
+    private Action parseDeleteAction(Object o) throws EVRException {
         if(o == null) {
             throw new EVRException("delete action cannot be null or empty");
         }
@@ -806,7 +608,7 @@ public class EVRParser {
         return action;
     }
 
-    private static List<String> parseDeleteLabelList(Object o) throws EVRException {
+    private List<String> parseDeleteLabelList(Object o) throws EVRException {
         List<String> labels = new ArrayList<>();
         if (o == null) {
             return labels;
@@ -821,7 +623,7 @@ public class EVRParser {
         return labels;
     }
 
-    private static List<GrantAction> parseDeleteAssociations(Object o) throws EVRException {
+    private List<GrantAction> parseDeleteAssociations(Object o) throws EVRException {
         List<GrantAction> associations = new ArrayList<>();
         if (o == null) {
             return associations;
@@ -836,7 +638,7 @@ public class EVRParser {
         return associations;
     }
 
-    private static List<EvrNode> parseDeleteNodes(Object o) throws EVRException {
+    private List<EvrNode> parseDeleteNodes(Object o) throws EVRException {
         List<EvrNode> nodes = new ArrayList<>();
         if (o == null) {
             return nodes;
