+++ src/main/java/gov/nist/csd/pm/pip/obligations/evr/EventParser.java
@@ -1,0 +1,13 @@
+package gov.nist.csd.pm.pip.obligations.evr;
+
+import gov.nist.csd.pm.pip.obligations.model.EventPattern;
+
+import java.util.Map;
+
+public interface EventParser {
+
+    String key();
+
+    EventPattern parse(Map map) throws EVRException;
+
+}
