+++ src/test/resources/epp/time.yml
@@ -1,0 +1,15 @@
+label: time test
+rules:
+  event:
+    time:
+      - "Mon 01/01/2020"
+      - "Wed 9:00:00"
+      - "Fri 01/01/2021 10:00:00"
+  response:
+    create:
+      - what:
+          name: newOA
+          type: OA
+        where:
+          name: oa1
+          type: OA
